Move deterministic rounding method to base class
