+++ src/aspire/abinitio/commonline_lud.py
@@ -5,8 +5,6 @@ from scipy.sparse import csr_array
 from scipy.sparse.linalg import eigs
 
 from aspire.abinitio import CLOrient3D
-from aspire.utils import nearest_rotations
-from aspire.utils.matlab_compat import stable_eigsh
 
 logger = logging.getLogger(__name__)
 
@@ -449,102 +447,3 @@ class CommonlineLUD(CLOrient3D):
         restructured_G = np.block([[top_left, top_right], [bottom_left, bottom_right]])
 
         return restructured_G
-
-    def _deterministic_rounding(self, gram):
-        """
-        Deterministic rounding procedure to recover the rotations from the Gram matrix.
-
-        The Gram matrix contains information about the first two columns of every rotation
-        matrix. These columns are extracted and used to form the remaining column of every
-        rotation matrix.
-
-        :param gram: A 2n_img x 2n_img Gram matrix.
-
-        :return: An n_img x 3 x 3 stack of rotation matrices.
-        """
-        logger.info("Recovering rotations from Gram matrix.")
-
-        # Obtain top eigenvectors from Gram matrix.
-        d, v = stable_eigsh(gram, 5)
-        sort_idx = np.argsort(-d)
-        logger.info(f"Top 5 eigenvalues from (rank-3) Gram matrix: {d[sort_idx]}")
-
-        # Only need the top 3 eigen-vectors.
-        v = v[:, sort_idx[:3]]
-
-        # According to the structure of the Gram matrix, the first `n_img` rows, denoted v1,
-        # correspond to the linear combination of the vectors R_{i}^{1}, i=1,...,K, that is of
-        # column 1 of all rotation matrices. Similarly, the second `n_img` rows of v,
-        # denoted v2, are linear combinations of R_{i}^{2}, i=1,...,K, that is, the second
-        # column of all rotation matrices.
-        v1 = v[: self.n_img].T
-        v2 = v[self.n_img : 2 * self.n_img].T
-
-        # Use a least-squares method to get A.T*A and a Cholesky decomposition to find A.
-        A = self._ATA_solver(v1, v2)
-
-        # Recover the rotations. The first two columns of all rotation
-        # matrices are given by unmixing V1 and V2 using A. The third
-        # column is the cross product of the first two.
-        r1 = np.dot(A.T, v1)
-        r2 = np.dot(A.T, v2)
-        r3 = np.cross(r1, r2, axis=0)
-        rotations = np.stack((r1.T, r2.T, r3.T), axis=-1)
-
-        # Make sure that we got rotations by enforcing R to be
-        # a rotation (in case the error is large)
-        rotations = nearest_rotations(rotations)
-
-        return rotations
-
-    @staticmethod
-    def _ATA_solver(v1, v2):
-        """
-        Uses a least squares method to solve for the linear transformation A
-        such that A*v1=R1 and A*v2=R2 correspond to the first and second columns
-        of a sequence of rotation matrices.
-
-        :param v1: 3 x n_img array corresponding to linear combinations of the first
-            columns of all rotation matrices.
-        :param v2: 3 x n_img array corresponding to linear combinations of the second
-            columns of all rotation matrices.
-
-        :return: 3x3 linear transformation mapping v1, v2 to first two columns of rotations.
-        """
-        # We look for a linear transformation (3 x 3 matrix) A such that
-        # A*v1'=R1 and A*v2=R2 are the columns of the rotations matrices.
-        # Therefore:
-        # v1 * A'*A v1' = 1
-        # v2 * A'*A v2' = 1
-        # v1 * A'*A v2' = 0
-        # These are 3*K linear equations for 9 matrix entries of A'*A
-        # Actually, there are only 6 unknown variables, because A'*A is symmetric.
-        # So we will truncate from 9 variables to 6 variables corresponding
-        # to the upper half of the matrix A'*A
-        n_img = v1.shape[-1]
-        truncated_equations = np.zeros((3 * n_img, 9), dtype=v1.dtype)
-        k = 0
-        for i in range(3):
-            for j in range(3):
-                truncated_equations[0::3, k] = v1[i] * v1[j]
-                truncated_equations[1::3, k] = v2[i] * v2[j]
-                truncated_equations[2::3, k] = v1[i] * v2[j]
-                k += 1
-
-        # b = [1 1 0 1 1 0 ...]' is the right hand side vector
-        b = np.ones(3 * n_img)
-        b[2::3] = 0
-
-        # Find the least squares approximation of A'*A in vector form
-        ATA_vec = np.linalg.lstsq(truncated_equations, b, rcond=None)[0]
-
-        # Construct the matrix A'*A from the vectorized matrix.
-        # Note, this is only the lower triangle of A'*A.
-        ATA = ATA_vec.reshape(3, 3)
-
-        # The Cholesky decomposition of A'*A gives A (lower triangle).
-        # Note, that `np.linalg.cholesky()` only uses the lower-triangular
-        # and diagonal elements of ATA.
-        A = np.linalg.cholesky(ATA)
-
-        return A
